requirements.txt
fastapi
uvicorn[standard]
python-dotenv
strands-agents[ollama]
strands-agents-tools